Speed up umc_plg_include
Instead of scraping a folder, a simple array could be provided to speed this up. It may be better just having the array in a register.php and include that separately to keep the code clean. It's the most simple approach I can think of. Just remember to register new plugins to the array.